# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for the AI Job Interview Platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Interviewer {
  id: number;
  name: string;
  title: string;
  company: string;
  personality: string;
  /** OpenAI TTS voice identifier */
  voiceId: string;
  /**
   * URL or base64 of the avatar image
   * @nullable
   */
  avatarUrl?: string | null;
}

export interface CreateSessionBody {
  /** The job role/title being interviewed for */
  jobRole: string;
  /**
   * Optional full job description text
   * @nullable
   */
  jobDescription?: string | null;
  /** Target interview duration in minutes (30-45) */
  durationMinutes?: number;
}

export interface Session {
  id: number;
  jobRole: string;
  /** @nullable */
  jobDescription?: string | null;
  durationMinutes: number;
  /** active | completed */
  status: string;
  /** IDs of selected interviewers for this session */
  interviewerIds: number[];
  currentInterviewerIndex: number;
  questionCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface QAItem {
  id: number;
  questionText: string;
  /** @nullable */
  answerText?: string | null;
  interviewerId: number;
  isFollowUp: boolean;
  questionIndex: number;
  createdAt: string;
}

export interface SessionWithQA {
  session: Session;
  questions: QAItem[];
  interviewers: Interviewer[];
}

export interface SubmitAnswerBody {
  /** ID of the question being answered */
  questionId: number;
  /** Transcribed text of the user's spoken answer */
  answerText: string;
}

export interface NextQuestionResponse {
  /** True if the interview is complete */
  done: boolean;
  question?: QAItem;
  /** The interviewer who will ask the next question */
  interviewerId?: number;
  sessionStatus: string;
}

export interface PostureAnalysisBody {
  /** Base64-encoded JPEG webcam snapshot */
  imageBase64: string;
}

export interface PostureAnalysisResult {
  /** Posture score 0-100 */
  score: number;
  /** Detailed posture feedback */
  feedback: string;
  /** List of identified posture issues */
  issues: string[];
  timestamp: string;
}

export interface AnswerFeedback {
  questionId: number;
  questionText: string;
  /** @nullable */
  answerText?: string | null;
  score: number;
  feedback: string;
  strengths: string[];
  improvements: string[];
}

export interface InterviewReport {
  sessionId: number;
  /** Overall interview score 0-100 */
  overallScore: number;
  communicationScore: number;
  technicalScore: number;
  confidenceScore: number;
  postureScore: number;
  answerFeedback: AnswerFeedback[];
  postureNotes: string[];
  /** Top improvement suggestions */
  suggestions: string[];
  /** Overall performance summary paragraph */
  summary: string;
  generatedAt: string;
}

export interface TtsRequestBody {
  /** Text to convert to speech */
  text: string;
  /** Interviewer ID to use their assigned voice */
  interviewerId: number;
}

export interface TtsResponse {
  /** Base64-encoded MP3 audio */
  audioBase64: string;
  format: string;
}

export interface TranscribeBody {
  /** Base64-encoded audio blob (webm/mp3/wav) */
  audioBase64: string;
  /** Audio MIME type */
  mimeType?: string;
}

export interface TranscribeResponse {
  /** Transcribed text */
  text: string;
}