# lib/api-zod/src/generated/types/answerFeedback.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for the AI Job Interview Platform
 * OpenAPI spec version: 0.1.0
 */

export interface AnswerFeedback {
  questionId: number;
  questionText: string;
  /** @nullable */
  answerText?: string | null;
  score: number;
  feedback: string;
  strengths: string[];
  improvements: string[];
}