# lib/api-zod/src/generated/types/interviewReport.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for the AI Job Interview Platform
 * OpenAPI spec version: 0.1.0
 */
import type { AnswerFeedback } from "./answerFeedback";

export interface InterviewReport {
  sessionId: number;
  /** Overall interview score 0-100 */
  overallScore: number;
  communicationScore: number;
  technicalScore: number;
  confidenceScore: number;
  postureScore: number;
  answerFeedback: AnswerFeedback[];
  postureNotes: string[];
  /** Top improvement suggestions */
  suggestions: string[];
  /** Overall performance summary paragraph */
  summary: string;
  generatedAt: Date;
}