# lib/api-zod/src/generated/types/interviewer.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for the AI Job Interview Platform
 * OpenAPI spec version: 0.1.0
 */

export interface Interviewer {
  id: number;
  name: string;
  title: string;
  company: string;
  personality: string;
  /** OpenAI TTS voice identifier */
  voiceId: string;
  /**
   * URL or base64 of the avatar image
   * @nullable
   */
  avatarUrl?: string | null;
}