# lib/api-zod/src/generated/types/nextQuestionResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for the AI Job Interview Platform
 * OpenAPI spec version: 0.1.0
 */
import type { QAItem } from "./qAItem";

export interface NextQuestionResponse {
  /** True if the interview is complete */
  done: boolean;
  question?: QAItem;
  /** The interviewer who will ask the next question */
  interviewerId?: number;
  sessionStatus: string;
}