# lib/api-zod/src/generated/types/postureAnalysisResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for the AI Job Interview Platform
 * OpenAPI spec version: 0.1.0
 */

export interface PostureAnalysisResult {
  /** Posture score 0-100 */
  score: number;
  /** Detailed posture feedback */
  feedback: string;
  /** List of identified posture issues */
  issues: string[];
  timestamp: Date;
}