# lib/api-zod/src/generated/types/sessionWithQA.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for the AI Job Interview Platform
 * OpenAPI spec version: 0.1.0
 */
import type { Interviewer } from "./interviewer";
import type { QAItem } from "./qAItem";
import type { Session } from "./session";

export interface SessionWithQA {
  session: Session;
  questions: QAItem[];
  interviewers: Interviewer[];
}