# lib/api-zod/src/generated/types/submitAnswerBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for the AI Job Interview Platform
 * OpenAPI spec version: 0.1.0
 */

export interface SubmitAnswerBody {
  /** ID of the question being answered */
  questionId: number;
  /** Transcribed text of the user's spoken answer */
  answerText: string;
}