# lib/api-zod/src/generated/types/transcribeBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for the AI Job Interview Platform
 * OpenAPI spec version: 0.1.0
 */

export interface TranscribeBody {
  /** Base64-encoded audio blob (webm/mp3/wav) */
  audioBase64: string;
  /** Audio MIME type */
  mimeType?: string;
}