# lib/api-zod/src/generated/types/ttsRequestBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for the AI Job Interview Platform
 * OpenAPI spec version: 0.1.0
 */

export interface TtsRequestBody {
  /** Text to convert to speech */
  text: string;
  /** Interviewer ID to use their assigned voice */
  interviewerId: number;
}