# lib/api-zod/src/generated/types/ttsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for the AI Job Interview Platform
 * OpenAPI spec version: 0.1.0
 */

export interface TtsResponse {
  /** Base64-encoded MP3 audio */
  audioBase64: string;
  format: string;
}